configure pyproject.toml
