+++ pyproject.toml
@@ -1,2 +1,32 @@
 [build-system]
-requires = ["pypandoc", "setuptools>=40.8.0", "wheel"]
+requires = ["setuptools>=40.8.0", "wheel"]
+
+[project]
+name = 'elog_zulip'
+authors = [{name = "Thomas Michelat", email = "thomas.michelat@gmail.com"},]
+description = "Publish elog entries to zulip"
+requires-python = ">=3.9"
+dynamic = ['version', 'readme']
+dependencies = [
+    'bs4',
+    'dataset',
+    'py-elog @ https://github.com/paulscherrerinstitute/py_elog/archive/refs/tags/1.3.16.tar.gz',
+    'html2text',
+    'jinja2',
+    'loguru',
+    'pandas',
+    'sqlalchemy==1.4.48',
+    'tabulate>=0.8.10',
+    'toml',
+    'zulip>=0.7.1',
+]
+
+[project.scripts]
+elog-zulip-publisher = 'elog_zulip:main'
+
+[tool.setuptools]
+packages = ['elog_zulip']
+
+[tool.setuptools.dynamic]
+version = {attr = 'elog_zulip.__version__'}
+readme = {file = ["README.md"]}
